+++ tests/test_callback_for.py
@@ -14,6 +14,11 @@ class FakeItemPage(ItemPage):
     def to_item(self):
         return 'fake item page'
 
+class FakeItemPageAsync(ItemPage):
+
+    async def to_item(self):
+        return 'fake item page'
+
 
 class FakeItemWebPage(ItemWebPage):
 
@@ -28,6 +33,12 @@ class MySpider(scrapy.Spider):
     parse_web = callback_for(FakeItemWebPage)
 
 
+class MySpiderAsync(scrapy.Spider):
+
+    name = 'my_spider_async'
+    parse_item = callback_for(FakeItemPageAsync)
+
+
 def test_callback_for():
     """Simple test case to ensure it works as expected."""
     cb = callback_for(FakeItemPage)
@@ -39,6 +50,20 @@ def test_callback_for():
     assert list(result) == ['fake item page']
 
 
+@pytest.mark.asyncio
+async def test_callback_for_async():
+    cb = callback_for(FakeItemPageAsync)
+    assert callable(cb)
+
+    fake_page = FakeItemPageAsync()
+    response = DummyResponse('http://example.com/')
+    result = cb(response=response, page=fake_page)
+
+    assert await result.__anext__() == 'fake item page'
+    with pytest.raises(StopAsyncIteration):
+        assert await result.__anext__()
+
+
 def test_callback_for_instance_method():
     spider = MySpider()
     response = DummyResponse('http://example.com/')
@@ -47,12 +72,16 @@ def test_callback_for_instance_method():
     assert list(result) == ['fake item page']
 
 
-def test_callback_for_inline():
-    callback = callback_for(FakeItemPage)
+@pytest.mark.asyncio
+async def test_callback_for_instance_method_async():
+    spider = MySpiderAsync()
     response = DummyResponse('http://example.com/')
-    fake_page = FakeItemPage()
-    result = callback(response, page=fake_page)
-    assert list(result) == ['fake item page']
+    fake_page = FakeItemPageAsync()
+    result = spider.parse_item(response, page=fake_page)
+
+    assert await result.__anext__() == 'fake item page'
+    with pytest.raises(StopAsyncIteration):
+        assert await result.__anext__()
 
 
 def test_default_callback():
@@ -85,6 +114,18 @@ def test_inline_callback():
     """Sample request with inline callback."""
     spider = MySpider()
     cb = callback_for(FakeItemPage)
+    request = scrapy.Request('http://example.com/', callback=cb)
+    with pytest.raises(ValueError) as exc:
+        request_to_dict(request, spider)
+
+    msg = f'Function {cb} is not an instance method in: {spider}'
+    assert str(exc.value) == msg
+
+
+def test_inline_callback_async():
+    """Sample request with inline callback using async callback_for."""
+    spider = MySpiderAsync()
+    cb = callback_for(FakeItemPageAsync)
     request = scrapy.Request('http://example.com/', callback=cb)
     with pytest.raises(ValueError) as exc:
         request_to_dict(request, spider)